backend/requirements.txt
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
python-decouple==3.8
celery==5.4.0
redis==5.2.1
anthropic==0.42.0
PyMuPDF==1.25.3
python-docx==1.1.2
gunicorn==23.0.0
